requirements.txt
asgiref==3.3.1
certifi==2020.12.5
dj-database-url==0.5.0
Django==3.1.4
django-background-tasks==1.2.5
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-compat==1.0.15
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
wincertstore==0.2
